adds docstrings, moves parse_tag to utils and adds documentation
